Got basic controls working.
Based off http://www.opengl-tutorial.org/beginners-tutorials/tutorial-6-keyboard-and-mouse/

This could be a lot better. It's not properly using the callbacks the
way the current glfw docs suggest we should. Also need to
rethink/refactor the test implementation, as we could pass in or provide
the Proj and View matrices. Then the test just needs to care about the
model.

Also would be nice to figure out if I really want mouse control. If not,
need to remove it completely and have a substitute for the view angle.